fix: updated table and migration

+++ budapp/playground_ops/models.py
@@ -133,7 +133,7 @@ class ChatSetting(Base, TimestampMixin):
     sequence_length: Mapped[int] = mapped_column(Integer, nullable=True)
     context_overflow_policy: Mapped[str] = mapped_column(String, nullable=False)  # enum
     stop_strings: Mapped[list[str]] = mapped_column(PG_ARRAY(String), nullable=True)
-    top_k_sampling: Mapped[float] = mapped_column(Integer, nullable=True)
+    top_k_sampling: Mapped[int] = mapped_column(Integer, nullable=True)
     repeat_penalty: Mapped[float] = mapped_column(Float, nullable=True)
     top_p_sampling: Mapped[float] = mapped_column(Float, nullable=True)
     min_p_sampling: Mapped[float] = mapped_column(Float, nullable=True)
